gradio_project/requirements.txt
gradio
langchain
openai
python-dotenv
exa_py
duckduckgo-search
llamaindex 